working on transactions front

+++ client/src/products/characteristics-mapper.tsx
@@ -1,14 +1,15 @@
-import { Characteristic } from "../product-types"
+import { Characteristic } from "./product-types"
 
 type Params = {
     characteristics: Array<Characteristic>,
-    onRemove: Function
+    onRemove?: Function
 }
 
 const CharacteristicsMapper = (params: Params) => {
     const {characteristics, onRemove} = params;
 
     const handleRemove = (index: number) => {
+        if(!onRemove) return;
         const temp = characteristics;
         temp.splice(index, 1);
         onRemove(temp);
@@ -24,11 +25,11 @@ const CharacteristicsMapper = (params: Params) => {
                         <div>{
                                 item.value
                             }</div>
-                        <div>
+                        {onRemove && <div>
                             <button type="button" onClick={() => handleRemove(
                                 characteristics.indexOf(item)
                             )}>видалити</button>
-                        </div>
+                        </div>}
                     </div>
             })
         }
